refactor: move files to utils and domain package
