+++ src/main/java/racingcar/Application.java
@@ -1,5 +1,7 @@
 package racingcar;
 
+import racingcar.utils.GameHandler;
+
 public class Application {
 	public static void main(String[] args) {
 		// TODO 구현 진행
