+++ src/main/java/racingcar/utils/UserService.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.utils;
 
 import java.util.HashSet;
 import java.util.Set;
